+++ SystemPaulinaVianaMachado/SystemPaulinaVianaMachado/src/main/java/com/grupoBlueSpectre/SystemPaulinaVianaMachado/Core/Usecases/Intefaces/AlterarUserCase.java
@@ -1,5 +1,5 @@
 package com.grupoBlueSpectre.SystemPaulinaVianaMachado.Core.Usecases.Intefaces;
 
 public interface AlterarUserCase<E> {
-    E execute(int ID, E e);
+    E execute(Long ID, E e);
 }
