+++ SystemPaulinaVianaMachado/SystemPaulinaVianaMachado/src/main/java/com/grupoBlueSpectre/SystemPaulinaVianaMachado/Core/Usecases/Intefaces/BuscarPorIDUserCase.java
@@ -3,5 +3,5 @@ package com.grupoBlueSpectre.SystemPaulinaVianaMachado.Core.Usecases.Intefaces;
 import java.util.Optional;
 
 public interface BuscarPorIDUserCase<E>{
-    Optional<E> execute(int ID);
+    Optional<E> execute(Long ID);
 }
